Add JUnit tests
