+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "java.project.sourcePaths": [
+        "src\\main\\java",
+        "src\\test\\java"
+    ]
+}
