+++ duke/data/duke.txt
@@ -1,0 +1,2 @@
+T###O###thing
+T###O###thing2
