+++ src/main/java/duke/ItemList.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.items.Item;
 import java.util.LinkedList;
 
 public class ItemList extends LinkedList<Item> {
