+++ src/main/java/duke/Storage.java
@@ -1,5 +1,9 @@
 package duke;
 
+import duke.items.Event;
+import duke.items.Deadline;
+import duke.items.Item;
+import duke.items.ToDo;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
