--- src/main/java/duke/ToDo.java
@@ -1,12 +1,0 @@
-package duke;
-
-public class ToDo extends Item {
-    public ToDo(String name) {
-        super(name);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
