+++ src/main/java/duke/items/ToDo.java
@@ -1,0 +1,12 @@
+package duke.items;
+
+public class ToDo extends Item {
+    public ToDo(String name) {
+        super(name);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
